+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 2.0.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Raise 404 if object no longer available
+  [vangheem]
 
 
 2.0.2 (2019-01-15)